[HUDI-5851] Improvement of data skipping, only converts expressions to evaluators once
