+++ hudi-flink-datasource/hudi-flink/src/main/java/org/apache/hudi/source/evaluator/Evaluator.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.hudi.source.evaluator;
+
+import org.apache.hudi.source.stats.ColumnStats;
+
+import java.io.Serializable;
+import java.util.Map;
+
+/**
+ * Decides whether it's possible to match based on the column values and column stats.
+ * The evaluator can be nested.
+ */
+public interface Evaluator extends Serializable {
+
+  /**
+   * Decides whether it's possible to match based on the column stats.
+   *
+   * @param columnStatsMap column statistics
+   * @return
+   */
+  boolean eval(Map<String, ColumnStats> columnStatsMap);
+}
